requirements.txt
rich>=13.0.0
ollama>=0.1.0
PyYAML>=6.0
psutil>=5.9.0